Squashed commit of the following:

commit a33aa5fe8e9b4dfbbabb31762a6e9994b9687e1e
Merge: d56b91a bb8e344
Author: tjuric <tereza.juric@infobip.com>
Date:   Wed Sep 20 09:25:11 2017 +0200

    Merge branch 'master' of https://git.ib-ci.com/scm/mml/infobip-mobile-messaging-android into tjuric-MM-2058-Manage_external_sdk_modules

commit d56b91a8b2cc7891661ac1f899091c9895e92cf3
Author: tjuric <tereza.juric@infobip.com>
Date:   Wed Sep 20 09:24:56 2017 +0200

    PR context fixes

commit 848237a858ff3336aee003c39787795f52c2b776
Author: tjuric <tereza.juric@infobip.com>
Date:   Tue Sep 19 22:12:38 2017 +0200

    Deleted unnecessary extended class

commit 0fea73efb1d9a46ddc24abb8fbb9192974ae4d85
Author: tjuric <tereza.juric@infobip.com>
Date:   Tue Sep 19 21:41:36 2017 +0200

    MM-2058 - More tests

commit 58e816c2df6f6e3ffa0e5a5950eb2fd5fb5b66ff
Author: tjuric <tereza.juric@infobip.com>
Date:   Tue Sep 19 19:16:03 2017 +0200

    Changes after merge

commit 38ae1f10269a8596603363aed301d4168a1d5dbf
Merge: 69873da e2298b8
Author: tjuric <tereza.juric@infobip.com>
Date:   Tue Sep 19 19:13:58 2017 +0200

    Merge remote-tracking branch 'remotes/origin/master' into tjuric-MM-2058-Manage_external_sdk_modules

    # Conflicts:
    #	infobip-mobile-messaging-android-interactive-sdk/src/main/java/org/infobip/mobile/messaging/interactive/notification/NotificationActionTapReceiver.java
    #	infobip-mobile-messaging-android-sdk/src/main/java/org/infobip/mobile/messaging/MobileMessagingCore.java

commit 69873da4c43d410283d45e319e8fe89f71444e64
Author: tjuric <tereza.juric@infobip.com>
Date:   Tue Sep 19 18:58:49 2017 +0200

    MM-2058 Tests

commit 70c3aa5bbed2c4490e4830bb4bb278bb74096caa
Author: tjuric <tereza.juric@infobip.com>
Date:   Tue Sep 19 17:17:32 2017 +0200

    MM-2058 Manage external modules

+++ infobip-mobile-messaging-android-interactive-sdk/src/androidTest/java/org/infobip/mobile/messaging/interactive/tools/MobileMessagingTestCase.java
@@ -131,7 +131,9 @@ public abstract class MobileMessagingTestCase extends MobileMessagingBaseTestCas
             } catch (Exception ignored) {
             }
         }
-        databaseProvider.deleteDatabase();
+
+        if (databaseProvider != null)
+            databaseProvider.deleteDatabase();
     }
 
     public static class TestMessageStore implements MessageStore {
@@ -187,6 +189,14 @@ public abstract class MobileMessagingTestCase extends MobileMessagingBaseTestCas
                 .withId(givenTappedActionId)
                 .withIcon(android.R.drawable.btn_default)
                 .withTitleResourceId(android.R.string.ok);
+    }
+
+    @NonNull
+    protected NotificationCategory givenNotificationCategory(NotificationAction tappedAction) {
+        return new NotificationCategory(
+                "categoryId",
+                tappedAction,
+                givenNotificationAction("actionIdNotTapped").build());
     }
 
     protected Intent givenIntent(Message message, NotificationCategory notificationCategory, NotificationAction action, int notificationId, int flags) {
